[Admin] Upload image, manange route

+++ resources/views/admin/company/index.blade.php
@@ -64,7 +64,7 @@
                         @endphp
                         @forelse ($companies as $company)
                             <tr>
-                                <td><input type="checkbox" value="{{ $company->id }}" name="companies_choice[]" class="choice-item flat-red"></td>
+                                <td><input type="checkbox" value="{{ $company->id }}" name="item_choice[]" class="choice-item flat-red"></td>
                                 <td>{{ $loop->iteration }}</td>
                                 <td><a href="{{ route('company.show', $company->id) }}">{{ $company->name }}</a></td>
                                 <td>{{ $company->address }}</td>
@@ -98,54 +98,48 @@
                     </tbody>
                 </table>
             </div>
-            <div class="box-footer clearfix">
-                <div class="input-group-btn">
-                    <button type="button" class="btn btn-warning btn-sm dropdown-toggle" data-toggle="dropdown" aria-expanded="true">Action
-                    <span class="fa fa-caret-down"></span></button>
-                    <ul class="dropdown-menu">
-                        <li>
-                            <a href="#" data-message="@lang('message.warning_delete_company')" class="btn-delete-multy">@lang('main.delete')</a>
-                            {{ Form::open(['method' => 'delete', 'url' => route('company.delete_multy'), 'class' => 'form-delete-multy']) }}
-                                {{ Form::hidden('data', '', ['class' => 'data-id-delete']) }}
-                            {{ Form::close() }}
-                        </li>
-                    </ul>
-                </div>
-                <div class="input-group-btn text-right">
-                    {{ Form::open(['method' => 'post', 'url' => route('company.update_multy_status'), 'class' => 'form-change-multy-status']) }}
-                        {{ Form::hidden('data', '', ['class' => 'data-change-status']) }}
-                        <button type="button"
-                            class="btn btn-info btn-sm btn-change-multy-status"
-                            data-message="@lang('message.warning_change_status_company')">
-                            @lang('company.change_status')
-                        </button>
-                    {{ Form::close() }}
-                </div>
-            </div>
-            <div class="box-footer clearfix">
-                <div class="col-md-3">
-                    <div class="dataTables_length" id="example1_length">
-                        <label>@lang('main.show')
-                            {{ Form::select('paginate', config('setting.number_of_records_pagination'), Request::get('size'), [
-                                'class' => 'show-number-record form-control input-sm'
-                            ]) }}
-                            @lang('main.entries')
-                        </label>
+            @if ($companies->total())
+                <div class="box-footer clearfix">
+                    <div class="input-group-btn">
+                        <button type="button" class="btn btn-warning btn-sm dropdown-toggle" data-toggle="dropdown" aria-expanded="true">Action
+                        <span class="fa fa-caret-down"></span></button>
+                        <ul class="dropdown-menu">
+                            <li>
+                                <a href="#" data-message="@lang('message.warning_delete_company')" class="btn-delete-multy">@lang('main.delete')</a>
+                                {{ Form::open(['method' => 'delete', 'url' => route('company.delete_multy'), 'class' => 'form-delete-multy']) }}
+                                    {{ Form::hidden('data', '', ['class' => 'data-id-delete']) }}
+                                {{ Form::close() }}
+                            </li>
+                        </ul>
+                    </div>
+                    <div class="input-group-btn text-right">
+                        {{ Form::open(['method' => 'post', 'url' => route('company.update_multy_status'), 'class' => 'form-change-multy-status']) }}
+                            {{ Form::hidden('data', '', ['class' => 'data-change-status']) }}
+                            <button type="button"
+                                class="btn btn-info btn-sm btn-change-multy-status"
+                                data-message="@lang('message.warning_change_status_company')">
+                                @lang('company.change_status')
+                            </button>
+                        {{ Form::close() }}
                     </div>
                 </div>
-                <div class="text-right col-md-9">
-                    {{ $companies->appends(Request::except('page'))->links() }}
+                <div class="box-footer clearfix">
+                    <div class="col-md-3">
+                        <div class="dataTables_length" id="example1_length">
+                            <label>@lang('main.show')
+                                {{ Form::select('paginate', config('setting.number_of_records_pagination'), Request::get('size'), [
+                                    'class' => 'show-number-record form-control input-sm'
+                                ]) }}
+                                @lang('main.entries')
+                            </label>
+                        </div>
+                    </div>
+                    <div class="text-right col-md-9">
+                        {{ $companies->appends(Request::except('page'))->links() }}
+                    </div>
                 </div>
-            </div>
+            @endif
         </div>
     </div>
 </div>
 @endsection
-
-@push('script')
-    <script>
-        $(document).ready(function() {
-            
-        })
-    </script>
-@endpush
